requirements.txt
Django==5.2.9
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery>=5.3
redis
django-resized
Pillow
django-phonenumber-field==8.4.0
gunicorn==23.0.0
python-dotenv==1.2.1
PyJWT==2.10.1
django-environ==0.12.0
drf-yasg==1.21.6
